feat(patient-import): shared validation between pd and dq
Ref: #1040
